PigCortisol(dataRep)	R Documentation

Diurnal rythm of cortisol in pigs on two different treatments

Description

In an experiment investigating the effect of stress on pigs (control vs. intermittent stress), the cortisol level in the blood was sampled every half hour from 12 to 36 hours after implantation of a catheter. Cortisol is a hormone involved in the stress response. The pigs were blocked according to litter.

Usage

data(PigCortisol)

Format

A data frame with 900 observations on the following 7 variables.

time

Time in hours post operation, a numeric vector

cortisol

The cortisol level, a numeric vector

sint

sin(time*2*pi/24), a numeric vector

cost

cos(time*2*pi/24), a numeric vector

pig

The ID of the pig, a factor with a level for each pig

litter

The Litter of the pig, a factor with levels 1,...,8

treat

The treatment of the pig, a factor with levels 1 2

Details

Source

Jensen, K.H., Pedersen, L.J., Heller, K.E., Ladewig, J. and Jørgensen, E. (1996). Intermittent Stress in pigs: effects on behavior, pituarity-adrenocortical axis, growth and gastric ulceration. Physiology & Behaviour, (59), p 741-748,

References

Examples


data(PigCortisol)
with(PigCortisol,plot(time, cortisol,log="y"))

[Package dataRep version 0.0-2 Index]