PigWeights(dataRep)	R Documentation

Weight and feed consumption of pigs

Description

Weight and feed consumption of 25 pigs on 5 different treatments in 4 replicates of (10,5,5,5) pigs. The replicates started at approximately weekly intervals

Usage

data(PigWeights)

Format

A data frame with 161 observations on the following 9 variables.

Group

a factor with levels 1 2 3 4 5, indicating the treatment the pig received

Pig

The identification of the pig

WeightNo

a numeric vector indicating the number of the weighing

Weight

The weight of the pig at the weighing, a numeric vector

ExpDays

Days since the experiment started, a numeric vector. Different replicates have different ExpDays when the pigs start in the experiment (Days==0)

FUp

Feed intake since last weighing (Feed Unit pig), a numeric vector

Gain

Weight Gain since last weighing, a numeric vector

AccFUp

Accumulated feed intake since the pig started in the experiment, anumeric vector

Days

Days since the pig started in the experiment, a numeric vector

Details

Source

The data are from an experiment conducted by Niels Oksbjerg

References

Examples



 library(lattice)
 data(PigWeights)
 with(PigWeights,xyplot(Weight~Days|Pig))

[Package dataRep version 0.0-2 Index]