activityCowsLong(dataRep)	R Documentation

Daily activity of cows.

Description

Twenty loggers placed on twenty cows transmit automatically daily activity measures to a PC. The data contain additional information about the farms on which the measurements were taken.

Usage

data(activityCowsLong)

Format

A dataframe with 7 variables.

logger:

id of logger

cow:

id of cow

farm:

id of farm

owner:

Owner of the farm

type:

Production type of the farm

Day:

Day of measurement

Activity:

Actitvity measurement

Source

References

See Also

activityFarmCow

Examples


data(activityCowsLong)  

[Package dataRep version 0.0-3 Index]