activityCowsWide(dataRep)	R Documentation

Daily activity of cows.

Description

Twenty loggers placed on twenty cows transmit automatically daily activity measures to a PC. The data contain additional information about the farms on which the measurements were taken. The data are in wide format as the three activity measurments for each cow are given in one row. For the long-format where the activity measurments are given in one variable see activityCowsLong

Usage

data(activityCowsWide)

Format

A dataframe with 8 variables.

logger:

id of logger

cow:

id of cow

farm:

id of farm

owner:

Owner of the farm

type:

Production type of the farm

day1:

activity measurement for day 1

day2:

activity measurement for day 2

day3:

activity measurement for day 3

Source

References

See Also

activityCowsLong

Examples


data(activityCowsWide)  

[Package dataRep version 0.0-3 Index]