anther(dataRep)	R Documentation

Embryogenic anthers.

Description

Anthers of the plant species Datura innoxia were prepared under different conditions and the number of embryogenic anthers counted.

Usage

data(anther)

Format

A dataframe with 4 variables.

y:

number of embryogentic anthers

n:

number of anthers prepared

storage:

storage conditions

force:

Centrifuging force

Source

References

Dobson, Annette, (2002), An Introduction to Generalized Linear Models, Chapma and Hall

Examples


data(anther)  

[Package dataRep version 0.0-3 Index]