bactsucrose(dataRep)	R Documentation

Influence of nutritients on growth of bacteria

Description

In an experiment on nutrients effects on bacterial growth, bacteria were grown on agar plates in solutions with different concentrations of the nutrients sucrose (4 levels) and leucine (3 levels). The experiment was replicated on 4 days and on each day all 12 treatment combinations were applied. The numbers of bacteria in the solution were determined.

Usage

data(bactsucrose)

Format

Data frame containing
day sucrose leucine density
day:

day of experiment

sucrose:

level of scurose nutrient

leucine:

level of leucine nutrient

density:

number fo bacteria

Details

Source

Grafen, A. and Hails, R. (2002), Modern Statistics for the Life Sciences, Oxford University Press, London, chapter 9.4, p. 169

References

Examples


data(bactsucrose)

[Package dataRep version 0.0-3 Index]