beets(dataRep)	R Documentation

Sugar beets data

Description

Yield of sugar beets as function of sowing and harvesting time in a split–plot experiment.

Usage

data(beets)

Format

A data frame with 30 observations on the following 5 variables.

harvest

Harvest time. A factor with levels harv1 harv2. The corresponding dates are 2/10 and 21/10.

block

The blocks, see details below.

sow

Sowing time. A factor with levels sow1 sow2 sow3 sow4 sow5. The corresponding dates are 4/4, 12/4, 21/4, 29/4 and 18/5

yield

Yield in Kg per 25 m2

sugpct

Sugar percentage

Details

The experiment was laid out as a as split plot experiment in blocks. There are three blocks each containing two whole–plot units corresponding to the harvest time. Each whole–plot unit is sub–divided into 5 split–plot units corresponding to the sowing dates.

Source

~~ reference to a publication or URL from which the data were obtained ~~

References

~~ possibly secondary sources and usages ~~

Examples


data(beets)
## maybe str(beets) ; plot(beets) ...

[Package dataRep version 0.0-2 Index]