bolger(dataRep)	R Documentation

Rodents in habitat fragments

Description

Prence/absence of any specy of native rodent (except Microtus califcornicus) dependent on distance to the nearest source canyon, age since the fragment was isolated by urbanization and the percentage of fragment area covered in shrubs (bushes).

Usage

data(bolger)

Format

This data frame contains:

pershrub:

percentage of fragment area covered in shrubs

distx:

distance [m] to the nearest source canyon

age:

age [years] since fragment was isolated by urbanization

rodentspa:

presence/absence of native rodent species [1=present, 0=absent]

References

Quinn, Gerry. P. and Keough, Michael J. (2002) Experimental Design and Data Analysis for Biologists, University Press, Cambridge Box 13-2, Fig 13-2

Examples


data(bolger)

[Package dataRep version 0.0-3 Index]