caesar(dataRep)	R Documentation

infection from birth by Caesarian section

Description

data collected from Weissenbacher, R.E, Klinikum Grosshadern, Munich, on the occurrence of infections of birth by Caesarian section.
logistic regression
sparse cells with non-existence of ML-estimator

Usage

data(caesar)

Format

A dataframe with 5 variables.

infect:

yes: infection after birth; no: no infection

risk:

1: risk factors as diabetics or excessive weight are present; 0: not present

antib:

1: antibiotics given as prophylaxe; 0: no antibiotics

caesp:

1: Caesarian section planned; 0: not planned

r:

number of births with infection

Source

Fahrmeyer, Tutz (2001) Multivariate statistical modellng based on Generalized Linear Models, Springer, 2nd edition, pages 3, 30 and 51;

Examples



data(caesar)
attach(caesar)

#printing the data as a flat contingency table:
  #  data.frame to contigency table
    f<-xtabs(r~infect+risk+antib+caesp)
  # flat contigency table
    f<-ftable(caesp+infect~antib+risk,data=f)

[Package dataRep version 0.0-2 Index]