cannibaleggs(dataRep)	R Documentation

Tribolium: cannibalistic egg eating.

Description

Study to analyse cannibalistic behavior of three species of the beetle of genus Tribolium. 150 eggs (50 of each species) were exposed to adult beetles of one species and it was recorded how many eggs of which species they ate.

Usage

data(cannibaleggs)

Format

A dataframe with variables.

adult:

adult Species: A=castaneum, b= confusum, C=madens

egg:

species of the egg

day:

indicator of the day the experiment was performed

y:

number of eggs out of 50 that were NOT eaten

Source

References

McCullagh, P. and Nelder, J.A.(1989), Generalized Linear Models, Chapman and Hall, Exercise 4.25, p. 143

Examples


data(cannibaleggs)  

[Package dataRep version 0.0-3 Index]