carclaims(dataRep)	R Documentation

Insurance claims for cars

Description

The data provide the average claims for damage of an owner's car (privately owned and comprehensively insured). Besides the average claim the number of claims in different classes are given

Usage

data(carclaims)

Format

Data frame containing

pa:

age group of police holder: 1=17-20, 2= 21-24, 3= 25-29; 4=30-34, 5=35-39, 6=40-49, 7=50-59, 8= 60-

cg:

car group; A-D

va:

car's age 100-3, 2= 4-7, 3= 8-9, 4=10-

y:

average claim in pounds

number:

number of claims

Details

Source

Mc Cullagh, P. and Nelder, J.A. Generalized Linear Models, Chapman and Hall, 1989, 2nd edition, p. 298

References

Baxter, L.A. Coutts, S.M. and Ross, G.A.F. Applications of linear models in motor insurance, Proceedings of the 21st International Congress of Actuaries, Zürich, pp. 11-29

Examples


data(carclaims)

[Package dataRep version 0.0-3 Index]