carrotfly(dataRep)	R Documentation

Damage of carrots by carrot fly larvae

Description

Wheatley and Freeman (1982) analysed an experiment on the impact of two different insecticides, applied in different soil depths, on the damage of carrots by carrot fly larvae. Additionally to the two insecticides two control experiments without applied insecticides were performed. The experiment was repeated three times.

From each plot on which a treatment had been applied a number of carrots were chosen, washed and classified as damaged or not.

Usage

data(carrotfly)

Format

Data frame containing

insecticide:

diazinon, disulfoton, control

depth:

depth the insecticide was applied [cm]

dam1:

number of damaged carrots in replicate 1

exam1:

number of examined carrots in replicate 1

dam2:

number of damaged carrots in replicate 2

exam2:

number of examined carrots in replicate 2

dam3:

number of damaged carrots in replicate 3

exam3:

number of examined carrots in replicate 3

Details

Source

Wheatley and Freeman (1982) A method of using the proportions of undamaged carrots or parsnips to estimate the relative population densities of carrot fly (Psila rosae) larvæ, and its practical applications, Annals of Applied Biology, p. 229-244

References

James W. Hardin, Joseph M. Hilbe,(2007) Generalized Linear Models and Extensions, 2nd edition, Stata Press, Texas, p. 79

Examples


data(carrotfly)
#reshape the data into long fromat
carrot<-reshape(carrotfly,direction='long',varying=list(c('dam1','dam2','dam3'),
                           c('exam1','exam2','exam3')),
                     v.names=c('dam','exam'),
                     times =c(1,2,3),timevar='replicate')
#removing the id variable
carrot<-subset(carrot,select=-id)

[Package dataRep version 0.0-3 Index]