catpills(dataRep)	R Documentation

Palatibility of pill.

Description

Four diffeerent formulation of a drug were tested for palatibility on 27 cats. On days 1,2,3,4,8,9,10 and 11 the degree of accept of the drug was recorded

Usage

data(catpills)

Format

A dataframe with variables.

catid:

cat id

sex

sex of cat, 1: make, 0: female

breed

breed of cat

weight

weight of a cat at the begin of the study

age

age of a cat at the begin of the study

fi

formulation of the drug at day i

yi

score at day i: 1=readily accepted, 2:accepted, 3: reluctantly accepted, 4: accepted in food, 5: refused

Source

References

Hand, David and Crowder, Martin (1996) Practical Longitudinal Data Analysis, Chapman and Hall, Example 10.1, page 146

Examples


data(catpills)  
catL<-reshape(catpills,direction='long',
     varying=list(paste('f',c(1:4,8:11),sep=''),
                  paste('y',c(1:4,8:11),sep='')),
     v.names=c('form','score'),timevar='day',
     idvar='catid')
catL$u<-ifelse(catL$form  <=3 ,'accept','NONaccept')




[Package dataRep version 0.0-3 Index]