cervical(dataRep)	R Documentation

Cervical cancer deaths

Description

The deaths from cervical cancer of women and the thousand women-years at risk are given for four age groups and four Eurpean countries for the year between 1969 and 1973.

Usage

data(cervical)

Format

This data frame contains:

country:

European country

age:

age group

deaths:

number of deaths from cervical cancer

womenyear:

1000 womenyear at risk, i.e the total time divided by 1000 that all women of a country belong to the corresponding age-group.

Details

Source

Whitttemore, A.S. and Gong G. (1991) Poisson regression with misclassified counts: application to cervical cancer mortality rates, JRSS, C, 40, 81-93

References

Lindsey, J.K. (2002) Modelling Frequency and Count Data, Oxford: Claredon Press, Exercise 7.10(3), p. 179

Examples


data(cervical)

[Package dataRep version 0.0-3 Index]