cheeseTaste(dataRep)	R Documentation

taste of four different cheese

Description

Persons were randomly assigned to four different types of cheeses and asked to characterize the taste on a scale from 1 to 9

Usage

data(cheeseTaste)

Format

A dataframe with 3 variables.

taste

response: evaluation of taste: from 1: strong dislike to 9: excellent taste

cheese:

type of cheese tasted

count:

number of answers

Source

McCullagh, P. and Nelder, J. A. (1989) Generalized Linear Models, London, U.K.: Chapman & Hall

[Package dataRep version 0.0-3 Index]