cherrytree(dataRep)	R Documentation

volum height and diameter of cherry trees

Description

measurements of the girth(=diam), height and volume of timber in 31 felled black cherry trees. Note that girth is the diameter of the tree (in inches) measured at 4 ft 6 in above the ground.

Usage

data(cherrytree)

Format

This data frame contains the following columns:

diam:

diameter of tree in inch [2.54 cm]

height:

height of tree in feet [30.48 cm]

vol:

volume of tree in cubic feet

Details

Source

Atkinson, A. C. (1982) Regression diagnostics, transformations and constructed variables (with discussion). Journal of the Royal Statistical Society, Series B, 44, 1-36.

References

Ryan, T. A. Jr., Joiner, B. L. and Ryan, B. F. (1985) The Minitab Student Handbook, Boston: Duxbury Press, 328-329.

Examples


data(cherrytree)
#

[Package dataRep version 0.0-2 Index]