cisplatin(dataRep)	R Documentation

reaction on chemotherapy with cisplatin

Description

The cisplatin data frame describe the severity of nausea for patients under chemotherapy, receiving cisplatinum or not.

Usage

data(cisplatin)

Format

This data frame contains the following columns:

cisplatinum:

no: patient received no cisplatinum, yes: received

severity:

level of serum acid phosphatase in King-Amstrong units

number

number of persons with the symptoms

Details

Source

Farewell(1982) A note on regression analysis of ordinal data with variability of classification, Biometrika, 69, 533-538

References

Lindsey, J. K (1995), Modelling Frequency and Count Data, Clarendon Press, Oxford, p. 99 ,reprint 2002

Examples


data(cisplatin)
#
library(MASS)

## Not run: 
* SAS example;
data prostaticCancer; 
infile "prostaticCancer.txt" firstobs=2;
input  age acid Xray Tsize  Tgrade  nodal;
proc print; run;
## End(Not run) 


[Package dataRep version 0.0-2 Index]