dentalgrowth(dataRep)	R Documentation

Dentalgrowth in children

Description

In a longitudinal study the dental growth in children was analysed. The distance from the center of the pituitary gland to the maxillary fissure for children at ages 8, 10, 12 and 14 was measured.

Usage

data(dentalgrowth)

Format

A data frame with 4 variables.

age

age of child when the distance was measured

dist

distance

id

identification number of child whithin sex

sex

sex of child

Details

Source

Myers, H.R. , Montgomery D.C., Vining, G.G. (2002) Generalized Linear Models, Exercise 6.6

References

~~ possibly secondary sources and usages ~~

Examples


data(dentalgrowth)
library(lattice)
trellis.device('windows',theme='col.whitebg')
xyplot(dist~age|sex,group=id,data=dentalgrowth,type=c('p','l'))

[Package dataRep version 0.0-2 Index]