devfun_vp {pbkrtest}	R Documentation

Compute deviance of a linear mixed model as a function of variance parameters

Description

This function is used for extracting the asymptotic variance-covariance matrix of the variance parameters.

Usage

devfun_vp(varpar, devfun, reml)

Arguments

varpar
variance parameters; varpar = c(theta, sigma).

devfun
deviance function as a function of theta only.

reml
if TRUE the REML deviance is computed; if FALSE, the ML deviance is computed.

Value

the REML or ML deviance.

Author(s)

Rune Haubo B. Christensen. Adapted to pbkrtest by Søren Højsgaard.

[Package pbkrtest version 0.5.2 Index]