dietox(dataRep)	R Documentation

Growth curves of pigs in a 3x3 factorial experiment

Description

The dietox data frame has 861 rows and 7 columns.

Usage

data(dietox)

Format

This data frame contains the following columns:

Weight

a numeric vector

Feed

a numeric vector

Time

a numeric vector

Pig

a numeric vector

Evit

a numeric vector

Cu

a numeric vector

Start

a numeric vector

Litter

a numeric vector

Source

Lauridsen, C., Højsgaard, S.,Sørensen, M.T. C. (1999) Influence of Dietary Rapeseed Oli, Vitamin E, and Copper on Performance and Antioxidant and Oxidative Status of Pigs. J. Anim. Sci.77:906-916

Examples


data(dietox)
str(dietox) 
plot(dietox)
## Not run: 
* SAS example;
proc print data=aaa;
run;
## End(Not run)


[Package dataRep version 0.0-3 Index]