dobson72Dis(dataRep)	R Documentation

Beetles killed by disulphide

Description

Same data as dobson72. Data is disaggregated such that for each individual beetle it is recorded whether it died ot not. Effect of gaseous carbon disulphide after 5 hours of exposure to beetles.
binary logistic regression

Usage

data(dobson72Dis)

Format

Data frame containing

dose:

dose of disulphide: [log10 of CS2 measured in mg/l]

y:

1: beetle was killed, 0: beetle survived

Details

Data are from the Table 7.2, page 119

Source

Dobson, A. (2002) An introduction to Generalized Linear Models, London: Chapman & Hall

Examples


data(dobson72Dis)

[Package dataRep version 0.0-3 Index]