eelworm(dataRep)	R Documentation

EElworm eelworms

Description

Experiment carried out the Rothamsted Experimental Station to to measure the effetiveness of 4 soil fimugants in keeping down the numner of eelworms in the soil.

Usage

data(eelworm)

Format

This data frame contains:

block:

Number of block, 1,2,3 or 4

plot:

Plot within block, 1,2,..., 12

conc:

Concentration of fumigant, 0= zero concentration, 1=base concentration, 2=twice the conetration as the base.

fumigant:

Fumigant used, CO=none (control), CN=chlorodinitrobenzene, CS=carbon disulphide jelly, CM=Cymag (propriety preparation), CK=Seekay (propriety preparation)

n.before:

Number of eelworm eelworms in a soil sample before application of a fumigant

n.after:

Number of eelworm eelworms in a soil sample after application of a fumigant

li:

The concentration of fumigant, conc, standardized, such that it has mean 0 and variance 1

qu:

The square of the concentration of fumigant, conc^2, standardized and made orthogonal to the variable li, such that sum(li*qu)=0 or said differently, the correlation cor(li,qu)=0.

treat:

Treatment generated from combining the 3 levels of concentration and the 4 fumigants. factor with 9 levels: 0.CO= no fumigant, 1.CK= concentration 1 and fumigant CK, 2.CK = concentration 2 and fumigant CK, etc.

Details

The experiment was laid out as a randomized block experiment, with 4 blocks and 12 plots whithin each block. The treatment combinations (concentration.fumigant) 0.CO, 0.CO, 0.CO, 0.CO, 1.CK, 2.CK, 1.CM, 2.CM, 1.CN, 2.CN, 1.CS, 2.CS were randomly applied to the 12 plots. The combination 0.CO was applied to 4 plots to ensure an accurate standard against which to compare the fumigants. ANOVA, ANCOVA, randomized block design

Source

Cochran W.G. and Cox, Gertrude M. (1957) Experimental Design chapters 3.2 and 3.8, Table 3.1, p. 46.

Examples

      
data(eelworm)

[Package dataRep version 0.0-3 Index]