esr(dataRep)	R Documentation

Determination of erythrocyte sedimenation rate(ESR)

Description

The ESR describes the rate at which the red blood cells (erythrocytes) settle out of a suspension of blood plasma. In an experiment at the Institute of Medical research, Kuala Lumpur, is was analyzed how the ESR is related to the plasma proteins fribinogen and gamma-globulin. An ESR<20 indicates a healthy individual.

Usage

data(esr)

Format

A dataframe

fib:

fibrinogen [gm/l]

glob:

gamma-globulin

esr:

0: esr<20; 1: 20<= esr

Source

Collet and Jemain(1985) Residuals, outliers and influential observations in regression analysis. Sains Malaysiana, 14, 493-511

References

Collet, D. (2003) Modelling Binary Data, 2nd edition, Chapman & Hall/CRC, London, Example 1.7

Examples


data(esr)
attach(esr)


[Package dataRep version 0.0-2 Index]