fatacid(dataRep)	R Documentation

Concentration of fish oil fatty acid x14 in tissue of pork

Description

A fish oil fatty acid x14 has been added in the concentrations in different concentrations to the food for pigs in a study. Interest is in seeing how much of the fatty acid can be found in the tissue.

Usage

data(fatacid)

Format

A data frame with 49 observations on the following 5 variables.

sample

a numeric vector

dose

a numeric vector

pigid

a numeric vector

litter

a numeric vector

x14

a numeric vector

Details

The pigs are fed with food containing the fatty acid. until their weight is 60 kg. From thereof and until they are slaughtered at 100kg, their food does not contain the fish oil. At 60kg (sample=1) and 100kg (sample=2) muscle biopsies are made and the concentration of x14 is determined.

Examples


data(fatacid)
## maybe str(fatacid) ; plot(fatacid) ...

[Package dataRep version 0.0-2 Index]