fatherson(dataRep)	R Documentation

Pearson's data on height of father and sons

Description

14 samples of a producte were randomly divided into 5 groups and each group was stored under different conditions. After storage the water content was measured.

Usage

data(fatherson)

Format

A dataframe with 2 variables.

father

height of father [inch]

son

height of son [inch]

Source

References

Examples


data(fatherson)
plot(father~son, data=fatherson)

[Package dataRep version 0.0-3 Index]