fertilhypo(dataRep)	R Documentation

Fertilhypo

Description

Constructed dataset about the yield of three fertilisers. There are three experiment. In the first the mean of the three fertilisers is 2, 5 and 9 and the residual variane is 0.36. In the iother tweo experiment the fertilisers have all mean 5, but the error variance is 5.29 and 0.36.

Usage

data(fertilhypo)

Format

This data frame contains:

fertil:

fertilosers 1,2 and 3

y1:

yield, mean per fertiliser is 2, 5 and 9, ero variance 0.36

y2:

yield, mean for all fertilisers 5, error varianece 5.29

y3:

yield, mean for all fertilisers 5, error varianece 0.36

Details

One way Anova, constructed data to show effect of different residual variances

Examples

      
data(fertilhypo)
#data generated by
set.seed(8)
treat  <-  rep(c(1,  2,  3),  each  =  c(10,  11,  12))
x  <-  1:length(treat)
treatf  <-  factor(treat)
y1  <-  c(2,  5,  9)[treatf]  +  rnorm(x,  sd  =  0.6)
y2  <-  5  +  rnorm(x,  sd  =  2.3)
y2  <-  ifelse(y2  <  0,  0.55,  y2)
y3  <-  5  +  rnorm(x,  sd  =  0.6)

[Package dataRep version 0.0-4 Index]