flagstone(dataRep)	R Documentation

flagstone

Description

Flagstones a priced fromconcrete. A manufacturer was interetses, wthther additives 'a', 'b' or a mixture of them, could imporve the stability of the flaghstone.

From a ppduciton of flagstones with the fo

Experiment to analyse the influence of meachanical stress on the growth of soya bean plants. 52 seedlings were potted, one seedling per plot. The seedlings were randomly splitted up into four groups of 13 plants. In two groups the plants were shaken for 20 minutes daily. The groups that received the same amount of stress were grown under two different light conditions. After 16 days the plants were harvested and the total area of leaves for each plant was measured.

Usage

data(flagstone)

Format

This data frame contains:

a:

Additive a: -1 not added, 1 : 10 kg of additive a added

b:

Additive b: -1 not added, 1 : 10 kg of additive a added

strength:

Compression strength of stone on maga pascal

Source

Blæsild, Preben and Grandfeldt, Jørgen (2003) Statistics with applications in Biology and Geology, Chapman and Hall, Exercise 4.8, pp. 239.

Examples

      
data(flagstone)

[Package dataRep version 0.0-3 Index]