flourBeetle(dataRep)	R Documentation

Toxicity of insecticides to flour beetles

Description

Flour beetles (tribolium castaneum) were sprayed with one of three different insecticides in solution in Shell oil P31.

The type of sprays were: DDT, gamma-BHC and a mixture of both.

The table contains the proportions of insects killed after a period of 6 days.

Usage

data(flourBeetle)

Format

A dataframe

killed:

beetles killed

total:

total number of beetles exposed

dose:

dosis of spray used [mg/10cm$^2$]

spray:

type of spray: ddt, bhc, mix

Source

Collet, D. (1991) Modelling Binary Data, Chapman & Hall, London, Example 4.3, 4.4, 5.9, p. 103

References

Examples


data(flourBeetle)
attach(flourBeetle)


[Package dataRep version 0.0-2 Index]