french(dataRep)	R Documentation

Regeneration and dispersal mechanism of plants

Description

In a study the number of occurrences between the regeneration mechanism of species (either by seed or only vegetatively) and their dispersal mechanism (by ants or vertebrate) was recorded.

Usage

data(french)

Format

This data frame contains:

regen:

regeneration mechanism, seed og veg(etatively)

dispers:

disperal mechanism, ant or vert(ebrate)

count:

number occurences in each class

Details

Log-linear model

Source

French, R.H and Westoby, M. (1996) vertebrate-dispersed species in fore-prone environment. Australian Journal of Ecology, 21, pp. 379–385

References

Quinn, Gerry. P. and Keough, Michael J. (2002) Experimental Design and Data Analysis for Biologists, University Press, Cambridge table 14.1, Box 14.2

Examples


data(french)
print(xtabs(count~regen+dispers,data=french))

[Package dataRep version 0.0-3 Index]