furnace(dataRep)	R Documentation

Recovery of heat

Description

The heat revovery of heat lost to the environment in the form of exhaust gases by two types of furnaces is analysed. For different settings of the speed of the exhaust gas passing the heat converter the rcovery ratio is measured.
Analysis of Covariance

Usage

data(furnace)

Format

A dataframe with 3 variables.

recovery:

ratio of heat recovered to heat lost

flow:

speed of exhaust gas [meter/second]

furnace:

type of furnace: A, B

Source

Milton, J. S. and Arnold, Jesse, C. (1995) Introduction to Probability and Statistics, McGrwa Hill, New York, p. 526, chapter 12, exercise 53

Examples



library(dataRep)
library(doBy)
data(furnace)

[Package dataRep version 0.0-2 Index]