gallbladder(dataRep)	R Documentation

Gallbladder volume of dogs

Description

Study on effect of chalagogues on changes in gallbladder volume of dogs. The experiment used as treatment two chalagogues and one control. Six dogs were assigned to each treatment. The gallbladder volumes were measured by ultrasound every ten minutes for two hours.

Usage

data(gallbladder)

Format

A data frame with following 4 variables.

treat:

8Clanobutin, Cholechystokynin or water as control

dogid:

dog indendtification number whithin each treatment

time:

time measured

gbv:

gallbladdervolume

Details

Source

Myers, H.R. , Montgomery D.C., Vining, G.G. (2002) Generalized Linear Models, Exercise 6.4

References

~~ possibly secondary sources and usages ~~

Examples


data(gallbladder)
library(lattice)
trellis.device('windows',theme='col.whitebg')
xyplot(gbv~time|treat,group=dogid,data=gallbladder,type=c('p','l'))

[Package dataRep version 0.0-2 Index]