get_Fstat_ddf {pbkrtest}	R Documentation

Compute denominator degrees of freedom for F-test

Description

From a vector of denominator degrees of freedom from independent t-statistics (nu), the denominator degrees of freedom for the corresponding F-test is computed.

Usage

get_Fstat_ddf(nu, tol = 1e-08)

Arguments

nu
vector of denominator degrees of freedom for the t-statistics

tol
tolerance on the consecutive differences between elements of nu to determine if mean(nu) should be returned

Details

Note that if any nu <= 2 then 2 is returned. Also, if all nu are within tol of each other the simple average of the nu-vector is returned. This is to avoid downward bias.

Value

the denominator degrees of freedom; a numerical scalar

Author(s)

Rune Haubo B. Christensen. Adapted to pbkrtest by Søren Højsgaard.

[Package pbkrtest version 0.5.2 Index]