get_covbeta {pbkrtest}	R Documentation

Compute covariance of fixed effect parameters as a function of variance parameters of a linear mixed model

Description

At the optimum the covariance is available as vcov(lmer-model). This function computes cov(beta) at non (RE)ML estimates of varpar.

Usage

get_covbeta(varpar, devfun)

Arguments

varpar
variance parameters; varpar = c(theta, sigma).

devfun
deviance function as a function of theta only.

Value

The covariances matrix of the fixed effects at supplied varpar values.

Author(s)

Rune Haubo B. Christensen. Adapted to pbkrtest by Søren Højsgaard.

[Package pbkrtest version 0.5.2 Index]