grainBeetle(dataRep)	R Documentation

Graint Beetles affected by fumnigant

Description

Grain Beetles were exposed to different doses of the fumigant ethylene oxide to test its effectiveness against insect pests.

The determination of whether the fumigant was affected could only be observed after several days after exposure and the beetles had to be transferred to glasstubes containing suitable food and having a temperature at about 25 degrees.

Usage

data(grainBeetle)

Format

A dataframe

conc:

concentration of ethylene oxide [mg/l]

affected:

number of beetles affected

exposed:

total number of beetles exposed

Source

Busvine, J.R.(1938) The toxicity of ethylene oxide to Calandra oryzae, C. granaria, Tribolium castaneum and Cimex lectularius, Annals of Applied Biology, 25, 605-632

References

Collet, D. (2003) Modelling Binary Data, 2nd edition, Chapman & Hall/CRC, London, Example 6.5; table 6.4

Examples


data(grainBeetle)
attach(grainBeetle)

[Package dataRep version 0.0-2 Index]