heparinclott(dataRep)	R Documentation

Clotting time of blood with heparin (HEPARINCLOTT)

Description

A standard preparation of heparin is compared to a test preparation by measuring the clotting time of blood for different doses of heparin. At each dose level two measurements are available.
normal, gamma with inverse link

Usage

data(heparinclott)

Format

A data frame with 20 observations and 3 variables.

time:

clotting time of blood (sec)

dose:

log-dose of heparin

preparation:

'standard' or 'test' preparation

Source

Osborn, J. F. (1979) Statistical Exercises in Medical Research, Blackwell, Oxford, p. 92, exercise 11.3.2

Examples


data(heparinclott)

[Package dataRep version 0.0-3 Index]