hodkin(dataRep)	R Documentation

Damage of carrots by carrotr fly larvae

Description

T4 cells/mm^3 in blood samples from 20 patients in remission from Hodgkin's disease and 20 patients in remission from disseminated malignancies.

Usage

data(hodkin)

Format

Data frame containing

hodk:

nuber of cells T4 per mm^3 from patients with Hodkin disease

nonhodk:

nuber of cells T4 per mm^3 from patients with Non-Hodkin disease

Details

Source

Altman, D. G. (1991) Practical Statistics for Medical Research, Chapman and Hall, London, p. 199

References

Lindsey, J.K. and Jones, B. (1998) Choosing among generalized linear models applied to medical data Statistics in Medicine, 17, p. 59-68

Examples


data(hodkin)
#reshape the data into long fromat
hodkinL<-reshape(hodkin,direction='long',varying=list(c('hodk','nonhodk')),v.names='ncell',times=c('yes','non'),timevar='hodkin')
hodkinL<-transform(hodkinL,hodkin=factor(hodkin))

[Package dataRep version 0.0-3 Index]