hubble(dataRep)	R Documentation

Hubble

Description

Hubbles data published in 1929 on the distance of a galaxy from ours and its velocity with which it moves away from the milkyway.

Usage

data(hubble)

Format

This data frame contains:

distance:

distance in 3.26 *1000000 light years

velocity:

km/second

References

Jens Ledet Jensen (2009) Et Nanokurs i Statistik, chapter 9, p. 96.

Examples

      
data(hubble)

[Package dataRep version 0.0-3 Index]