hyperinsuline(dataRep)	R Documentation

obesity and hyperinsulinism

Description

The data are from an experiment analysing the relation between obesity and hyperinsulinism. The plasma inorganic phosphate levels one hour after a standard glucose tolerance test were obtained for hyperinsulinemic obese and nemon-obese patients and a control group.

Usage

data(hyperinsuline)

Format

This data frame contains the following columns:

state:

yes: hyperinsulinemic obese patient; no: hyperinsulinemic non-obese patient; control: controls

phos:

level of phosphate levels [mg/dl] one hour after glucose tolerance test

rep:

number of repeated laboratory measurements of the same plasma sample

Details

The original data from Jones (1987) do not have the information on repetitive laboratory determinations. They were only included for exercise pupose.

Source

Jones, R.H. (1987) Serial correlations in unbalanced mixed models. Bulletin International Statistical Institute 52, book4, 105-122

References

Dobson, A. (1990) An introduction to Generalized Linear Models, London: Chapman & Hall, exercise 7.1

Examples


require(dataRep)
data(hyperinsuline)
#

[Package dataRep version 0.0-3 Index]