idiopathic(dataRep)	R Documentation

Treatment of idiopathic respiratory disease

Description

In a comparative trial of trishydroxymethyl-amino-methane (THAM) and sodium bicarbonate in the treatment of idiopathic respiratory distress syndrome 25 children were treated with THAM and 25 with sodum bicarbonate. It was observed whether the child survived or not. The disease is an acute lung disease of the newborn (especially the premature newborn). The lungs cannot expand because of a wetting agent is lacking.
binary logistic regression

Usage

data(idiopathic)

Format

A data frame with 50 observations on the following 5 variables.

child

id number of child

y

survival: y='survived' child survived, y='died' child died

treat

0: child received sodium bicarbonate, 1: chld got THAM

time

time of onset of spontaneous respiration [minutes]

weight

birthweight of the child [kg]

Source

Osborn, J. F. (1979) Statistical Exercises in Medical Research, Blackwell, Oxford, p. 87, exercise 10.4.2

Examples


data(idiopathic)

[Package dataRep version 0.0-2 Index]