infantLoss(dataRep)	R Documentation

Problemtic children and stillbirths

Description

In a study for children classified by their teachers as problematic (cases) and unproblematic(controls) the birth order of the child and the number of stillbirths (births with infant death) were recorded. The study was performed as a retrospective study, i.e. one chose first the number of problematic/non-problematic children and then collected the additional information.

Usage

data(infantLoss)

Format

Data frame containing

order:

birth order of the child, e.g. 1-2: first or second child

loss:

stillbirth: yes=a child died at birth

state:

control or case (=child classified)

n:

number of children

Details

Source

Cochran (1959) Some methods for strengthening the common chi^2 tests, Biometrics 10, 417-451

References

Fienberg, Stephen f. (1980) The Analysis of Cross-Classified Categorical Data, MIT Press, Cambridge, 2nd edition. p. 136 Table 7-7

Examples


data(infantLoss)

[Package dataRep version 0.0-3 Index]