kalk(dataRep)	R Documentation

Kalk

Description

Experiment about the effectiveness of lime as a diet supplement. 10 men received a diet with added lime, 11 without. Their blood pressure was measured at the begin of the study and 12 week after.

Usage

data(kalk)

Format

This data frame contains:

treatment:

kalk (=lime added) or placebo

before:

blood presssure at start

after:

blood presssure 12 weeks later

diff:

blood presssure before - after

Details

Anova, paired

Source

Blood prssure and metabolic effects of calcium supplementation in normotensive white and black men, JAMA, 1987, p. 1722-1776

References

Jens Ledet Jensen (2009) Et Nanokurs i Statistik, chapter 7, p. 71.

Examples

      
data(kalk)

[Package dataRep version 0.0-3 Index]