koagulat(dataRep)	R Documentation

koagulat

Description

The status of immogubolin was measured with a glutaraldehyd coagulation test. In one herad the results of the test was recored for heifers [dansk=kvier] and cows.

Usage

data(koagulat)

Format

This data frame contains:

koag:

results of the coagulation test at three levels, complete, incomplete and negative

type:

type of animal, heifer or cow

count:

the number of animals found in each category

Details

Rang test (Wilcoxon)

Source

Blom, J.. (1981) Enzootisk pneumoni hos kalve. Epidemiologi og profylakse. Licentatiatsafhandling, Institut for intern medicin, KVL Købehabvn

References

Søren Andersen (1989) Statistiske Analyse af Tælledata, Matematisk Institut, KVL p. 52, Kapitel 8, Eks. 1

Examples


data(koagulat)

[Package dataRep version 0.0-3 Index]