kyphosis(dataRep)	R Documentation

kyphosis in children after laminectomy surgery

Description

Kyphosis is the forward flexion of the spine of at least 40 degrees from vertical following a corrective spinal surgery.

For 83 childen potential risk factor as age, starting vertrebrae level of the surgery and number of vetrebrae levels were recorded.

Usage

data(kyphosis)

Format

A dataframe with 4 variables.

kyphosis:

absent, present

age:

age of patient in months

number:

number of vertebrae levels involved

start:

starting level of vertebrae level of the surgery

Source

Bell, D., Walker, J., O'Connor, G., Orrel, J., Tibshirani, R. (1989) Spinal deformation following multi-level thoracic and lumbar laminectomy in children

References

Hasti, T.J, Tibshirani, R.J. (1990) Generalized Additive Models, Chapman and Hall : London

Examples


data(kyphosis)
attach(kyphosis)

[Package dataRep version 0.0-3 Index]