leukemirat(dataRep)	R Documentation

Leukemi treatments in rats

Description

In an experiment 30 rats had a leukemic condition induced. Three chemotherapy drugs were applied and as a response the number of cancer cell colonies was observed. As covariates the number of red and white blood cells were collected. Each animal was observed at 4 times.

Usage

data(leukemirat)

Format

A data frame with 6 variables.

ratid:

rat identification number

drug:

number identifying which drug used

time:

time period when measurements were taken

wbc:

count of white blood cells

rbc:

count of red blood cells

y:

count of cell colonies

Details

Source

Myers, H.R. , Montgomery D.C., Vining, G.G. (2002) Generalized Linear Models, Exercise 6.6

References

~~ possibly secondary sources and usages ~~

Examples


data(leukemirat)
leukemirat$drug<-factor(leukemirat$drug)
library(lattice)
trellis.device('windows',theme='col.whitebg')
xyplot(y~time|drug,group=ratid,data=leukemirat,type=c('p','l'),span=1,
       strip=strip.custom(var.name='Drug',style=3,strip.name=TRUE))


[Package dataRep version 0.0-2 Index]