lizard(dataRep)	R Documentation

Lizard

Description

The data contain information about the habitat of two lizard species A. grahamii and A. opalinus of the genus Anolis. The questions is whether the lizards habitat is dependent on ecological factors describing the place (perch) where they were observed.

Usage

data(lizard)

Format

This data frame contains:

n:

number of lizards

sun:

Sun (=sunshine), Shade ( under shade)

height:

heightnif perch (High: >= 4.75 feet; Low: < 4.75 feet)

perch:

diameter of perch (Broad: > 4.0 inch; Narrow < 4.0 inch)

time:

time of observation (Morning, Mid.day, Afternoon)

species:

opalinus, grahamii

Details

log-linear model, two-way or higher way tables

Source

Schoener, T.W. (1968) The anolis lizards of Bimini: resource partitioning in a complex fauna, Ecology, 49, pp. 704–726

References

Crawley, Michael (2007). The R book, Wiley, p. 560

Examples


data(lizard)

[Package dataRep version 0.0-3 Index]