loyn(dataRep)	R Documentation

Abundance of birds

Description

Data contain for 56 forest patches in southeastern Victora, Australia, the abundance of forest birds in each patch, the distance of a patch to the neigbour and the nearest larger patch, grazing intensity by livestock, altitude of the patch and years since the isolation of the patch.

Usage

data(loyn)

Format

This data frame contains:

abund:

abundance of forest birds in patch

area:

patch area [ha]

yr.isol:

number of years since patch was isolated

dist:

distance to nearest patch [km]

ldist

distance to nearest larger patch [km]

graze:

grazing intensity in patch (1 to 5 indicating light to heavy)

alt:

altitude ([m] above sealevel) of patch

Details

GAM

Source

Loyn, R.H (1987) Effetcs of patch area and habitat on bird abundance, species numbers and tree health in fragmented Victorian forests, In: Nature Conservation: The role of remnants of native vegetation (Saunders, D.A.- Arnold, G.W., Burbidge, A.A, Hopkins, A.J.M. eds.) pp 65-77, Surrey Beatty and Sons, Chipping Norton, NSW

References

Quinn, Gerry. P. and Keough, Michael J. (2002) Experimental Design and Data Analysis for Biologists, University Press, Cambridge Box 13-3, Example for GAMs

Examples


data(loyn)

[Package dataRep version 0.0-3 Index]