minerswheeze(dataRep)	R Documentation

Breathing problems of coal miners

Description

For for coal miners aged between 20 and 64 it was recorded whether they suffered from breathlessness and whether they sufferend from wheezing, i.e. difficulty in breathing usually with a whistling sound.
logistic model, log linear model with two binary and one multicategrorical variable

Usage

data(minerswheeze)

Format

A dataframe with 3 variables.

age:

9 age groups= 20-24, 25-29, ...,60-64

breathless:

yes or no

wheeze:

wheezing: yues or no

number:

persons observed

Source

Ashford, J.R. and Sowden, R.R (1970) Multivariate probit analysis, Biometrics, 26, 535-546

References

Krzanowski, W. (1998) An Introduction to Statistical Modelling, Arnold, London, p. 213, Table 7.3

Examples


library(dataRep)
data(minerswheeze)

[Package dataRep version 0.0-2 Index]