nematodes(dataRep)	R Documentation

Number of nematodes

Description

Number of nematodes in 15 samples of solutions of 20 μl and 40 μl.

Usage

data(nematodes)

Format

This data frame contains

volume:

volume of solution μl

count:

number of nematodes counted in a sample solution

freq:

frequency of observations of the specific number of nematodes

Source

Lars Stubsgaard, Institute of Biological Sciences, University of Arhus

References

Blæsild, P. and Granfeldt, J. (2003) Statistics with applications in Biplogy and Geography, Chapman and Hall-CRC: London, p. 358

Examples


data(nematodes)


[Package dataRep version 0.0-3 Index]