orobanche(dataRep)	R Documentation

germination of orobache aegyptica

Description

For two varieties of seeds from Orobanche (parasitic plant growing on the roots of flowering plants) their ability to germinate being exposed to root extract frrom either bean or cumcumber was analysed. For each factor combination several plates were prepared with a root extract and seeded. \ a 2x2 factorial experiment \ two types of seed: O. aegyptiaco 73 (O73) O. aegyptiaco 75 (O75) \ two types of root extract bean (b) cucumber (c)

Usage

data(orobanche)

Format

A dataframe with 4 variables.

extract:

bean; cucumber

seed:

o73; o75

gem:

number seed germinated

total:

total number of seeds per plate

Source

Cox & Snell (1989) Analysis of Binary data, Chapman and Hall: London, Example 3.2 p. 114

Examples


data(orobanche)
attach(orobanche)


[Package dataRep version 0.0-3 Index]