otitis(dataRep)	R Documentation

Treatment of otitis

Description

Children who had one or both ears infected with otitis were either treated with the antibiotic cefaclor (CEF) or amoxicillin (AMO). After 14 days the state of their ears were reevaluated.

Usage

data(otitis)

Format

This data frame contains:

treat:

antibiotic treatment: CEF or AMO

entry:

number of ears with otitits at entry into the study

age:

age of child

nclear:

number of ears clear of acute otitis

nchild:

number of children

Source

Rosner (1989) Mutivariate methods for clustered binary data with more than one level of nesting JASA, 88, 373-380, Table 1.

References

Dobson, A. (2002) An introduction to Generalized Linear Models, London: Chapman & Hall, 2nd edition, Exercise 11.3, p. 211

Examples


data(otitis)

[Package dataRep version 0.0-3 Index]