pistons(dataRep)	R Documentation

Failure of piston rings

Description

These data concern the number of failures of piston rings for each of three legs of four steam-driven compressors.
ANOVA, Randomised complete block design, normal or poisson

Usage

data(pistons)

Format

Data frame containing

pistons:

Number of failures of piston rings

leg:

Position of leg (North=1, Centre=2, South=3)

compressor:

Compressor identification number

Source

Davies, O.L. and Goldsmith, P.L. (1972) Statistical Methods in Research and Production, 4th edn. Oliver and Boyd, Edinburgh, p. 324.

References

Examples


data(pistons)

[Package dataRep version 0.0-2 Index]