plaque(dataRep)	R Documentation

Effect of apple and orange eating on plaque formation

Description

In a trial the effect of eating of apples or oranges on the formation on dental plaque was analysed. Each child ate an apple after dinner the one week and the other week an orange. All children ate the same school meal and the sequence of apple and orange eaten was randomized. The extent of plaque was maesured with the Loe index defined as 0: no plaque 1: plaque invisible to the eye, but detactable by probes 2: moderate plaque visible 3: heavy deposits of plaque

Usage

data(plaque)

Format

A data frame with 30 observations on the following 3 variables.

child

id number of child

apple

Loe index for eating apples

age

Loe index for eating oanges

Source

Osborn, J. F. (1979) Statistical Exercises in Medical Research, Blackwell, Oxford, p. 106, exercoe 13.2.2

Examples


data(plaque)

[Package dataRep version 0.0-2 Index]