poisoned(dataRep)	R Documentation

Survial of poisoned animals

Description

The survival of animals exposed to threee different poisons and four treatments is analysed.

Usage

data(poisoned)

Format

A data frame with 48 observations on 3 variables.

time

survival time [10hr]

treat

type of treatment: A, B, C or D

poison

type of poison: 1,2 or 3

Source

Box, G. E. P. and Cox, D. R. (1964), An Analysis of Transformations, Journal of the Royal Statistical Society, 211-243, discussion 244-252.

Examples


data(poisoned)

[Package dataRep version 0.0-3 Index]