policestop(dataRep)	R Documentation

Stop by police.

Description

Over a 15 month period in 1998-1999 the police in New York recorded the ethneticity of and the type of crime of the stopped person

Usage

data(policestop)

Format

A dataframe with the variables.

stops:

number of persons stopped

pop:

size of the population

past.arrests:

number of persons arrested in the year before

precinct:

id number of the precinct

eth:

ethnicity: 1=black, 2=hispanic, 3=white

crime:

crime type: 1=violent, 2=weapons, 3=property, 4=drug

Source

References

Gelman, Andrew and Hill, Jennifer (2007), Data Analysis Using regression and Multilevel Hierrarchical Models

Examples


data(policestop)  

[Package dataRep version 0.0-3 Index]