primulaMendel(dataRep)	R Documentation

Mendel's law and primulæ

Description

Are some simple primula features consistent with simple Mendelian law

Usage

data(primulaMendel)

Format

Data frame containing

leaf:

type of leaf of a primula classified

eye:

type of eye of a classified primula

n:

number of plants classified

Details

A total of 560 plants were classified by type of their leafs (flat or crimped) and the type of their eye (normal or Primrose Queen). According to simple Mendelian law one would expect a 3:1 ratio for the dominant to the recessive type of each characteristic. For leafs 420 plants would be expected to have flat leaves and 140 plants crimped, and for the eye 420 normal and 140 Primrose Queen.

Source

Fisher, R. A. (1958) Statistical Methods for Research Workers. Edinburgh, Oliver and Boyd. p. 82

References

Lindsey, J.K (1995) Modelling Frequency and Count Data, Oxford University Press, p. 52, table 2.12. Mead. R, Curnow R.N., Hasted, A.M. (1993) Statistical Methods in Agriculture and Experimental Biology, end edition, Chapman and Hall: London, chapter 14.2, p. 302

Examples


data(primulaMendel)
attach(primulaMendel)

[Package dataRep version 0.0-3 Index]