protein(dataRep)	R Documentation

Diets with different contents of protein

Description

Six groups, each of ten rats, were fed on diets which differed according to source of protein and amount of protein in diet. The weight gain for each rat was recorded.
ANOVA, 2-way,

Usage

data(protein)

Format

Data frame containing

weight:

Weight gain [g]

protein:

Protein source: beef, cereal, pork

amount:

Amount of protein: low, high

Source

Snedecor, G.W. and Cochran, G.C. (1967) Statistical Methods, 6th edn. Iowa State University Press, p. 347.

Examples


data(protein)

[Package dataRep version 0.0-3 Index]