ratReproduction(dataRep)	R Documentation

Reproductivity of Rats

Description

Impact on diets on reproductivity of rats

Usage

data(ratReproduction)

Format

This data frame contains the following columns:

group:

treated, control

nsurv:

number of pups (out of 'ntotal') suriving the 21-day lactation period

ntotal

number of pups alive four days after birth

Details

Female rats were fed during their pregnancy two different diets.

The treatment group of 16 females rats received a diet with a chemical added.

The control group of 16 rats were fed without the chemical added.

After the birth of the litters, the number of pups that survived the 21-day lactation period, out of those that were alive four days after birth, was recorded.

Object of the experiment: Has the chemical an effect on the survivalof the pups.

Source

Weil, C.S. (1970) Selection of the valid number of sampling units and a consideration of their combination in toxicological studies involving reproduction, teratogenesis or carcigonesis, Food and Cosmetic Toxicology, 8, 177-182

References

Collet, D. (2003) Modelling Binary Data, 2nd edtition, Chapman & Hall, London, Example 1.4, p. 4

Examples



[Package dataRep version 0.0-2 Index]