ratsbodyweight(dataRep)	R Documentation

Development of body weight of rats.

Description

Two treatments were compared to a control on their effect on the growth of rats.

Usage

data(ratsbodyweight)

Format

A dataframe with 6 variables. ratid treat week0 week1 week2 week3 week4

ratid

id of rat

treat

treatment

week0

weight at start of experiment

week1

weight at week 1

week2

weight at week 2

week3

weight at week 3

week4

weight at week 4

Source

Box, G.E.P. (1950) Problems in the Analysis of Growth and Wear Curves, Biometrics, 6(4), 362-389

References

Examples


library(lattice)
data(ratsbodyweight)
rat<-reshape(ratsbodyweight,direction='long',varying=list(paste('week',c(0:4),sep='')),v.name='weight',idvar='ratid',timevar='week',time=c(0:4))
 print(
xyplot(weight~week|treat,groups=ratid,data=rat,type='l')
 )  

[Package dataRep version 0.0-3 Index]