rens(dataRep)	R Documentation

rens

Description

Preference between a new washing powder 'new' and a standard washing powder 'old'. Both washing powders were sent to 1008 consumers who were asked which powder they preferred, at which temperature they washed, the hardness of their water and whether they formerly had used the standard washing powder 'old'.

Usage

data(rens)

Format

This data frame contains:

hardness:

hardness of water: soft, medium, hard

pref:

preference of washing powder 'new' the consumer preffered 'new' to 'old', 'old': ther consumer preferred 'old' to 'new'.

consumer:

yes: the consumr had formerly usd the 'old' powder, no: the consumer never used the 'old' powder before

temp:

washing temperature: high, low

count:

number of consumers

Details

logistic regression, log-linear model, four-ways table

Source

Ries, P. N and Smith, H. (1963) The use of chi-square for preference testing in multidimensional problems, Chemical engineering progress, 59, 39-43

References

Søren Andersen (1989) Statistiske Analyse af Tælledata, Matematisk Institut, KVL p. 33, Kapitel 4, Eks. 3

Examples


data(rens)

[Package dataRep version 0.0-3 Index]