rlsfit(RLS)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


rlsfit(xx, yy, tvar=xx, degree = 1, lambda = 0.99,  b=NULL, Pi=NULL)

Arguments

xx 	 ~~Describe xx here~~

yy 	 ~~Describe yy here~~

tvar 	 ~~Describe yy here~~

b 	 ~~Describe b here~~

Pi 	 ~~Describe Pi here~~

lambda 	 ~~Describe lam here~~

degree 	 ~~Describe degree here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (xx, yy, b, Pi, lam = 0.5, degree = 1) 
{
    lenx <- length(xx)
    J <- diag(1, degree + 1)
    if (degree == 0) {
        X <- matrix(1, nrow = lenx, ncol = 1)
    }
    else {
        if (degree == 1) {
            X <- cbind(1, xx)
        }
        else {
            X <- cbind(1, xx, xx^2)
        }
    }
    if (missing(Pi)) 
        Pi <- diag(100, degree + 1)
    if (missing(b)) 
        b <- rep(0, degree + 1)
    ans1 <- matrix(NA, nrow = lenx, ncol = length(b))
    ans2 <- matrix(NA, nrow = lenx, ncol = 4)
    colnames(ans2) <- c("yt1.hat", "yt.hat", "et1", "et")
    for (tt in 1:lenx) {
        xt <- X[tt, ]
        yt <- yy[tt]
        if (!is.na(yt)) {
            k <- (Pi %*% xt/lam)/as.numeric(1 + t(xt) %*% Pi %*% 
                xt/lam)
            yt1.hat <- as.numeric(t(xt) %*% b)
            et1 <- yt - yt1.hat
            b <- b + k * et1
            Pi <- (J - k %*% t(xt)) %*% Pi/lam
            yt.hat <- as.numeric(t(xt) %*% b)
            et <- yt - yt.hat
            ans1[tt, ] <- b
            ans2[tt, ] <- c(yt1.hat, yt.hat, et1, et)
        }
    }
    ans <- cbind(ans1, ans2)
    ans
  }

[Package RLS version 0.1 Index]