roberts(dataRep)	R Documentation

Dead coolbah trees

Description

The presence of dead coolibah trees (Eucalyptus coolibah) was observed in sampled quadrats along a transect on floodplain. Of interest was the dependence of the presence of these trees on the position on the transect, either close to the dunes, to the lakeshore or intermediate.

Usage

data(roberts)

Format

This data frame contains:

coolbah:

presence of dead coolbah trees in sampling quadrat (with, without)

position:

position of the quadrat on the transect line (top=dunes, bottom=lokeshore, middle= intermediate)

count:

number of dead coolbah trees in the sampling quadrat

Details

logistic regression, Log-linear model

Source

Roberts, J. (1993) Regeneration and growth of coolibah Eucalyptus Coolabah subsp. arida, a riparian [dansk: bred] tree in the Cooper Crek region of South Australia, Australian Journal of Ecology, 18, pp. 345–350

References

Quinn, Gerry. P. and Keough, Michael J. (2002) Experimental Design and Data Analysis for Biologists, University Press, Cambridge Table 14-1 and Box 14-3

Examples


data(roberts)
print(xtabs(count~position+coolbah,data=roberts))

[Package dataRep version 0.0-3 Index]