seizlong(dataRep)	R Documentation

Epiliptic Seizures

Description

The dataset has the number of epiliptic seizures in each of four two-week intervals, and in a baseline eight-week inverval, for treatment and control groups with a total of 59 individuals.

Usage

data(seizlong)

Format

A data frame with 295 observations on the following 7 variables.

trt

a numeric vector, 0: placebo, 1: treatment

age

age of subject

time

time of registration, 0: prior to randomization, 1,2,3,4: follow-up visit

y

Number of epileptic seizures. For time 0: In the 8 week period prior to randomization. For time 1..4: In the 2 week period before each follow-up visit

id

Identification of subject

D

Variable given the length (2 or 8 weeks) of the registration period

x

a numeric vector. 0: prior to randomization, 1: after randomization

Source

Thall, P.F. and Vail S.C. (1990) Some covariance models for longitudinal count data with overdispersion. Biometrics *46*: 657-671.

References

Diggle, P.J., Liang, K.Y., and Zeger, S.L. (1994) Analysis of Longitudinal Data. Clarendon Press.

Examples


data(seizlong)
## maybe str(seizlong) ; plot(seizlong) ...

[Package dataRep version 0.0-2 Index]