shepdog(dataRep)	R Documentation

Shepdogs

Description

The hips of several shepdogs were x-rayed at the age of 1 year and the the degree of the disease dysplasie [dansk: hofeledsdysplasi] was evaluated on a scale of scores from 1 (=best) to 10 (=worst). The offspring of 5 fathers are contained in the data.

Usage

data(shepdog)

Format

This data frame contains:

father:

name of the father

score:

degree of disease: 1= lowest disease status, 10=severest disease status

count:

number of offspring

Details

log-linear model, two-ways table

Source

Røgen, E. (1985) 2nd HD statistik for 1984. Schæferhunden, 13, 22-24

References

Søren Andersen (1989) Statistiske Analyse af Tælledata, Matematisk Institut, KVL

Examples


data(shepdog)

[Package dataRep version 0.0-3 Index]