shuttleOrings(dataRep)	R Documentation

Failure of O-rings for shuttle flights

Description

The failure of gas sealing O-rings caused the explosion of the Challanger-shutttle shortly after launch.

Usage

data(shuttleOrings)

Format

Data frame containing

nr:

number of flight

date:

date of flight

n:

number of damage of the (primary) o-field rings out of 6 possible

temp:

average temperature (Fahrenheit)

pres:

Pressure applied to a gas for testing the sealing quaility of the rings

Details

On January, 28, 1986, the shuttle Challanger (flight 61-I) exploded shortly after the start. Later it was found that 'a combustion gas leak' caused the catastrophe. The leakage may by attributed to the failure of one or several of six O-rings, which task was to seal the field joints of the motor.

The night before the event engineers discussed the problem in a three hour telephone conference. In the conference a plot of the failure of O-rings in former missions were discussed, but only data for flights with failures were used, the flights without failures were considered not to contribute valuable data. Later the Rogers Commission noted that this was a failure of a proper data analysis.

Source

Dalal, S.R., Fowlkes, E.D. and Hoadley, B. Risk Analysis of the Space Shuttle: Pre-Challanger Prediction of Future, JASA, Vol. 84, p. 945-957

References

Examples


data(shuttleOrings)

[Package dataRep version 0.0-3 Index]