sideeffect(dataRep)	R Documentation

Sideffects of cancer operatiion

Description

Four different operation schemes for cancer of the duodenum [dansk= tolfingertarm kræft] were compared with respect to the side effects. The operation schemes were in different degree of the reduction of the stomach. From operation type 1 to 4 the amount of reduction increased.

Usage

data(sideeffect)

Format

This data frame contains:

operation:

type of operation, 1: small part of the stomach was removed, 4: large part of the stomach was removed

effect:

degree side effect

count:

number patients

Details

log-linear model, two-way table

Source

Grizzle, J. E., Starmer, C.F and Koch, G.G. (1969) Analysis of categorical data by linear models, Biometrics, 40, pp. 393–408

References

Søren Andersen (1989) Statistiske Analyse af Tælledata, Matematisk Institut, KVL

Examples


data(sideeffect)

[Package dataRep version 0.0-3 Index]