sinclair(dataRep)	R Documentation

Wildebeest [gnus]

Description

Wildbeest [dansk=gnu] carcasses from the Serengeti were cross-classified by sex, cause of death and and bone marrow type. The bone marrow type solid white fatty indicates a healthy animal not undernourished.

Usage

data(sinclair)

Format

This data frame contains:

sex:

male, female

marrow:

swf=solid white fatty, og= opaque gelatinous, tg= translucent gelatinous

death:

pred: killed by predator, npred: non-predator death cause

count:

number of wilbeests

Details

Logistic regression, log-linear model, three-way table

Source

Sinclair, A.R.E and Arcese, P. (1995) Population consequences of of predation-sensitive foraging: the Serengeti wildebeest [gnu], Ecology, 76, pp. 882–891

References

Quinn, Gerry. P. and Keough, Michael J. (2002) Experimental Design and Data Analysis for Biologists, University Press, Cambridge Table 14-3 and Box 14-5

Examples


data(sinclair)
tab<-xtabs(count~death+sex+marrow,data=sinclair)
print(ftable(tab))

[Package dataRep version 0.0-3 Index]