sitka89(dataRep)	R Documentation

Growth of Sitka Spruce Trees

Description

Impact of ozone on the growth of sitka spruce trees.

Usage

data(sitka89)

Format

A dataframe

size:

normal-bracket8bracket-normal
size of the tree measured in size of the tree measured in log(height*diamter^2)
time:

days after the 1st january, 1988

tree:

id number of a tree

treat:

ozone: grown under ozone environment, control: ozone free

Source

References

Examples


data(sitka89)
attach(sitka89)

[Package dataRep version 0.0-2 Index]