soilErosion(dataRep)	R Documentation

SoilErosion

Description

Experiment on the difference in erosion for two grass varieties.

Usage

data(soilErosion)

Format

This data frame contains:

grass:

grass variety: wheatgrass or cheatgrass

open:

log(length[inch]) of largest opening of bar soil

cover:

logit of the percentage of soil coverage.

erosion:

soil lost by erosion[pounds/acre]

Details

A 200-acre field was divided into 6X6 foot large subplots, their slope varying between 33 to 66 percent. The whole area was exposed to 1.8 inches of artificial rain. 13 of the finally used test-plots were covered with wheatgrass the other 16 plots with cheatgrass. The plots differed in the degree of soil coverage, which was characterized by the largest opening of bar soil [given in log-inches] and the percentage of ground cover [expressed as logit(p)=log(frac{p}{1-p}) of the percentage p]. The erosion was measured as the soil lost in pounds per acre.

References

Bliss, C. I. (1970) Statistics in Biology, Volume II, Mc Graw Hill, Düsseldorf, p. 323–326, Table 18.13

Examples

      
data(soilErosion)

[Package dataRep version 0.0-3 Index]