sowingmachine(dataRep)	R Documentation

Sowingmachine

Description

10 different fields were partitions into two subfields each. On one subfield a traditional sawing machine was used on the other a more modern machine. The harvest yields of the two machines were compared.

Usage

data(sowingmachine)

Format

This data frame contains:

fieldid:

id number of field

new:

yield using modern machine

old:

yield using old machine

Details

Anova, paired

Source

Wishart, J.(1934) Statistics in Agricultural Research Suppl. Journal Roy. Stat. Society

References

Jens Ledet Jensen (2009) Et Nanokurs i Statistik, chapter 7, p. 77.

Examples

      
data(sowingmachine)

[Package dataRep version 0.0-3 Index]