taulman(dataRep)	R Documentation

Squirrels

Description

The demographic distribution of caught flying squirrels in reponse to experimental logging (harvest) and year of the eperiment.

Usage

data(taulman)

Format

This data frame contains:

treat:

treatment (hervest, control

year:

year if experiment

age:

age of the caught flying squirrel

count:

number of squirrels caught

Details

Logistic regression, log-linear model, three-way table

Source

Taulman. J.F, Smith, K.G. and Thill, R.E. (1998) Demographic and behavioral responses of southern flying squirrels to experimental logging in Arkansas Ecological Applications, 8, pp. 1144-1155

References

Quinn, Gerry. P. and Keough, Michael J. (2002) Experimental Design and Data Analysis for Biologists, University Press, Cambridge Table 14-3 and Box 14-6

Examples


data(taulman)
tab<-xtabs(count~treat+year+age,data=taulman)
print(ftable(tab))

[Package dataRep version 0.0-3 Index]