timber(dataRep)	R Documentation

hardness and density of timber wood

Description

the janka hardness index and the density recorded for 36 Australian hard wood samples

The hardness is a difficult to measure quantity, and one wants to use density to predict the hardness.

Usage

data(timber)

Format

A dataframe with 2 variables.

hardness:

janka hardness index

density:

density of the wood

Source

Williams, E.J (1959) Regression Analysis, Wiley: New York

References

Venables, W.N. (2000) Exegeses on Linear Models paper Presented on the S-Plus User-s conference, Washington, DC, 8-9th October 1998 http://www.stats.ox.ac.uk/pub/MASS3/Exegeses.pdf

Examples


data(timber)
attach(timber)
par(mfrow=c(2,1))
plot(density,hardness)
fit<-lm(hardness~density+I(density^2))
plot(predict(fit),residuals(fit))
abline(h=0)
title('residuals of hardness~ density + density^2 vs. fitted values')

[Package dataRep version 0.0-3 Index]