titanium(dataRep)	R Documentation

extinction coefficient and titanium content of cement

Description

Experiment on the relation between the titanium content of cement and the extinction coefficient of solutions of the cement

Usage

data(titanium)

Format

A dataframe with 3 variables.

contents:

per cent titanium contents of the cement

length:

length of cuvette[cm]

extinction:

extinction coefficient

Source

Hald, A. (1952) Statistical Theory with Engineering Applications, New York: Wiley

References

Examples


data(titanium)
attach(titanium)


[Package dataRep version 0.0-2 Index]