tomat(dataRep)	R Documentation

Germination of tomatoes

Description

In an experiment to analyse the effect of watering on the germination of seeds, five watering regimes were applied to seeds on ten trays for each regime. On three successive dates the percentage of germinating seeds were measured.

Usage

data(tomat)

Format

Data frame containing

day:

number of day of measurement

water:

watering level, -2,-1,0

tray:

number of tray whtinin a watering regime

germ:

percentage of germinated seeds

Details

Source

Grafen, A. and Hails, R. (2002), Modern Statistics for the Life Sciences, Oxford University Press, London, chapter 8.4, p. 149

References

Examples


data(tomat)
tomat<-transform(tomat,datef=factor(date),trayf=factor(tray),waterf=factor(water))
library(lattice)
print(
xyplot(germ~date|waterf,groups=tray, type='l',data=tomat,
strip = strip.custom ( var.name='Water', strip.names=c(TRUE,TRUE),style=3),layout=c(5,1))
)


[Package dataRep version 0.0-3 Index]