trout(dataRep)	R Documentation

Weight and length of rainbow trout

Description

A data set with the length and weight of rainbow trouts of two sexes measured repeatedly over a 6 month period

Usage

data(fishw)

Format

A data frame with 2999 observations on the following 8 variables.

family.id

a numeric vector

animal.id

a numeric vector

vessel

a numeric vector

date

a numeric vector

length

a numeric vector

weight

a numeric vector

sex

a factor with levels f m

month

a numeric vector

Details

'fishwsub' is a subset of 'fishw'

Source

Data are collected by Majbrit Holst Nielsen, Danish Institute of Agricultural Sciences in the Autumn 2003

Examples


data(fishw)
## maybe str(fishw) ; plot(fishw) ...

[Package dataRep version 0.0-2 Index]