turtle(dataRep)	R Documentation

Hatching of turtle eggs

Description

The turtle data frame gives the number of male and female hatched eggs of turtles under different temperatures.

Usage

data(turtle)

Format

This data frame contains the following columns:

batch:

number of batch

temp:

temperature of the batches under hatching

male:

number of eggs with male turtles

female:

number of eggs with female turtles

state:

location whre the eggs were collected

Details

Source

K. Koehler, Iowa State Univeristy, see:
http://courses.ncssm.edu/math/Stat_Inst/PDFS/REG3_LOG.pdf

References

Examples


#data(turtle)
#

[Package dataRep version 0.0-3 Index]