vaso(dataRep)	R Documentation

Vaso

Description

The data an experiment to study the influence of the volume and the rate of air inspired by a human on the occurrence of transient vaso-constriction in the skin of the fingers. Four specific temperatures and two specific humidities were analysed. Vaso-constriction is the process by which blood vessels in the extremities (arms and legs) are constricted, preventing blood flow to certain tissues. Hence, on cold days fingers seem thinner.
To way ANOVA

Usage

data(vaso)

Format

A dataframe with 3 variables.

volume:

k

rate:

4 levels

constr:

binary: 1: constrriction observed, 0: no constriction

Source

Finney, D.J (1947) The estimation from original records of the relationship between dose and quantal response, Biometrika, 34, 320-334

References

Krzanowski, Wojtek J. (1998) An Introduction to Statistical Modelling Arnold, London

Examples


library(dataRep)
library(doBy)
library(lattice)
data (vaso)

[Package dataRep version 0.0-3 Index]