waistloss(dataRep)	R Documentation

Effect on weigth loss by participatin on a 'waist loss' program

Description

The weights of 20 men were recorded before and after participation in a 'waist loss' program

Usage

data(waistloss)

Format

Data frame containing

id:

id number of participant

before:

weight before start of program

after:

weight at the end of the program

Details

Data are from the exercise 2.5-2.2 of A. Dobson

Source

Dobson, A. (2002) An introduction to Generalized Linear Models, London: Chapman & Hall, 2nd edition

References

Examples


data(waistloss)
attach(waistloss)
diff.weight<-after-before
boxplot(diff.weight)
abline(h=0,col=2)
title('weight difference after-before')

[Package dataRep version 0.0-3 Index]