wandern(dataRep)	R Documentation

wandern

Description

365 walkers were asked how frequently they had a walk in the wood and how long the tours were on average.

Usage

data(wandern)

Format

This data frame contains:

distance:

distance of walk

frequency:

weekly: each week, monthly: each month, season: sveral times in season

count:

number of walkers

Details

log-linear model, two-ways table

Source

Haakenstad (1975) Skogbehandling i utfartsomåde. En opinionsundersøgelse om Oslomarkar, Meldinger fra Norges Landbrukshøgskole, Nr. 16

References

Søren Andersen (1989) Statistiske Analyse af Tælledata, Matematisk Institut, KVL p. 1, Eks. 1

Examples


data(wandern)

[Package dataRep version 0.0-3 Index]