watercontent(dataRep)	R Documentation

Water content change under different storing conditions

Description

14 samples of a producte were randomly divided into 5 groups and each group was stored under different conditions. After storage the water content was measured.

Usage

data(watercontent)

Format

A dataframe with 3 variables.

group

number of group

water

water content in per cent

Source

Hald, A. (1952) Statistical Theory with Engineering Applications, New York: Wiley, p. 432

References

Examples


data(watercontent)
attach(watercontent)


[Package dataRep version 0.0-3 Index]