get_Fstat_ddf {pbkrtest}	R Documentation

Compute denominator df for F-test

Description

From a vector of denominator df from independent t-statistics (nu), the denominator df for the corresponding F-test is computed.

Usage

get_Fstat_ddf(nu, tol = 1e-08)

Arguments

nu
vector of denominator df for the t-statistics

tol
tolerance on the consequtive differences between elements of nu to determine if mean(nu) should be returned

Details

Note that if any nu <= 2 then 2 is returned. Also, if all nu are within tol of each other the simple average of the nu-vector is returned. This is to avoid downward bias.

Value

the denominator df; a numerical scalar

Author(s)

Rune Haubo B. Christensen. Adapted to pbkrtest by Søren Højsgaard.

[Package pbkrtest version 0.5.1 Index]