get_Lb_ddf {pbkrtest}	R Documentation

Adjusted denomintor degress freedom for linear estimate for linear mixed model.

Description

Get adjusted denomintor degress freedom for testing Lb=0 in a linear mixed model where L is a restriction matrix.

Usage

get_Lb_ddf(object, L)
Lb_ddf(L, V0, Vadj)

Arguments

object
A linear mixed model object.

L
A vector with the same length as fixef(object) or a matrix with the same number of columns as the length of fixef(object)

V0, Vadj
Unadjusted and adjusted covariance matrix for the fixed effects parameters. Undjusted covariance matrix is obtained with vcov() and adjusted with vcovAdj().

Value

Adjusted degrees of freedom (adjusment made by a Kenward-Roger approximation).

Author(s)

Soren Hojsgaard, sorenh@math.aau.dk

References

Ulrich Halekoh, Søren Højsgaard (2014)., A Kenward-Roger Approximation and Parametric Bootstrap Methods for Tests in Linear Mixed Models - The R Package pbkrtest., Journal of Statistical Software, 58(10), 1-30., http://www.jstatsoft.org/v59/i09/

See Also

KRmodcomp, vcovAdj, model2restrictionMatrix, restrictionMatrix2model

Examples

(fmLarge <- lmer(Reaction ~ Days + (Days|Subject), sleepstudy))
## removing Days
(fmSmall <- lmer(Reaction ~ 1 + (Days|Subject), sleepstudy))
anova(fmLarge,fmSmall)

KRmodcomp(fmLarge,fmSmall)  ## 17 denominator df's
get_Lb_ddf(fmLarge, c(0,1)) ## 17 denominator df's

# Notice: The restriction matrix L corresponding to the test above
# can be found with
L<-model2restrictionMatrix(fmLarge, fmSmall)
L

[Package pbkrtest version 0.4-5 Index]